Rename event and game test files so they get run
